requirements.txt
# Core Streamlit and LangChain dependencies
streamlit>=1.28.0
langchain>=0.0.350
langchain-google-genai>=1.0.0

# Google Generative AI
google-generativeai>=0.3.0

# Document processing
pypdf>=3.17.0
python-docx>=0.8.11
unstructured>=0.11.0

# Vector database
faiss-cpu>=1.7.4

# Environment management
python-dotenv>=1.0.0

# Additional utilities
numpy>=1.24.0
pandas>=2.0.0

# Optional: For better document processing
# unstructured[local-inference]  # Uncomment for better document parsing